requirements.txt
adapter-transformers
bidict
datasets
evaluate
emoji
hydra-core
hydra-optuna-sweeper
jupyter
matplotlib
mlflow
nbformat
nltk
numpy
omegaconf
optuna
pandas
plotly
pytorch_lightning
scikit-learn
seaborn
sentence-transformers
seqeval
tensorboardX
torchtext
tqdm
transformers
